pyproject.toml
[project]
name = "resnet-demo"
version = "0.1.0"
description = "Demo TensorFlow model that learns to output x + 1 for inputs in [0, 255]."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "tensorflow-cpu==2.16.1",
]

[tool.uv]
dev-dependencies = []
